Move getDate to shared
